ToDoList_backend/requirements.txt
Django==4.2.21
djangorestframework==3.16.0
django-cors-headers==4.7.0
gunicorn
whitenoise